requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.63
botocore==1.35.63
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
deepface==0.0.93
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.3
fastapi-cli==0.0.5
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.67.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.2
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
